src - Move outdated logic to legacy folder
