+++ legacy/zone_raster.py
