+++ src/swed_17/__init__.py
@@ -1,9 +1,7 @@
 from .zone_compare import ZoneCompare
 from .zone_plotter import ZonePlotter
-from .zone_raster import ZoneRaster
 
 __all__ = [
     "ZoneCompare",
     "ZonePlotter",
-    "ZoneRaster",
 ]
